db: create utils
